update: split AwsCargoDependency into separate package rename: runtimeCrate to smithyRuntimeCrate remove: smithy prefix from RuntimeType
